+++ resources/js/domains/tickets/pages/Show.vue
@@ -2,18 +2,18 @@
     <div v-if="ticket">
         <h1><b>Details Ticket</b></h1><br>
         <h2><b>Titel:</b> {{ ticket.title }}</h2><br>
-        <p><b>Gebruiker:</b> {{ ticket.user }}</p><br>
+        <p><b>Gebruiker:</b> {{ ticket.user_name }}</p><br>
         <p><b>Beschrijving:</b><br>{{ ticket.content }}</p><br>
-        <p><b>Categorie:</b> {{ ticket.category }}</p><br>
+        <p><b>Categorie:</b> {{ ticket.category_name }}</p><br>
         <p><b>Status:</b> {{ formatStatus(ticket.status) }}</p><br>
-        <p><b>Toegewezen aan:</b> {{ ticket.assigned ? ticket.assigned : 'Nog niet toegewezen' }}</p><br>
+        <p><b>Toegewezen aan:</b> {{ ticket.assigned_name ? ticket.assigned_name : 'Nog niet toegewezen' }}</p><br>
         <p><b>Laatst gewijzigd:</b> {{ formatDate(ticket.updated_at) }}</p><br>
         <p><b>Reacties:</b></p>
         <div v-if="!ticket.reactions.length">
             <p>Er zijn nog geen reacties geplaatst.</p><br>
         </div>
         <div v-for="reaction in ticket.reactions" :key="reaction.id">
-            <p>{{ reaction.user }}:</p>
+            <p style="font-style: italic;">{{ reaction.user_name }}:</p>
             <p>{{ reaction.content }}</p>
             <br>
         </div>
@@ -29,7 +29,7 @@
 import { useRoute, useRouter } from 'vue-router';
 import { deleteTicket, fetchTicket, formatStatus, getTicketById } from '../store';
 import { ref } from 'vue';
-import { formatDate } from '../../../services/helpers';
+import { formatDate } from '../../../services/helpers/date';
 import { isCurrentUserAdmin } from '../../auth/store';
 
 const route = useRoute();
@@ -41,6 +41,8 @@ fetchTicket(ticketId);
 const ticket = ref(getTicketById(ticketId));
 
 const handleDelete = async () => {
+    const confirmation = confirm('Weet je zeker dat je deze ticket wilt verwijderen?');
+    if (!confirmation) return;
     await deleteTicket(ticket.value.id);
     router.push({ name: 'tickets.overview' });
 };
